Use 8 byte ints. Add printing code.

+++ advanced_solver_cpp/util.h
@@ -11,23 +11,23 @@ using std::min;
 class Util {
   Util() {}
 
-  static int getNumNeeded(int node, map<int, int>& inventory, map<int, int>& numRuns, const bool getMax, Problem& p) {
-    int totalNumNeeded = 0;
+  static int64_t getNumNeeded(int node, map<int, int64_t>& inventory, map<int, int64_t>& numRuns, const bool getMax, Problem& p) {
+    int64_t totalNumNeeded = 0;
     for (auto parent : p.inverseDependencies.at(node)) {
-      const int parentNumRuns = getNumRuns(parent, inventory, numRuns, getMax, p);
-      const int childPerParent = p.dependencies.at(parent).at(node);
+      const int64_t parentNumRuns = getNumRuns(parent, inventory, numRuns, getMax, p);
+      const int64_t childPerParent = p.dependencies.at(parent).at(node);
       const Fraction pq = max(Fraction(1), p.materialBonus.at(parent) * childPerParent);
 
-      int numNeededByParent = 0;
+      int64_t numNeededByParent = 0;
       if (getMax) {
-        int maxSlots = 0;
+        int64_t maxSlots = 0;
         if (p.job2machine.at(parent) == p.job2machine.at(node)) {
           maxSlots = min(parentNumRuns, (p.maxNumBatches.at(p.job2machine.at(parent)) - 1) * p.maxNumSlotsOfJob.at(parent));
         } else {
           maxSlots = min(parentNumRuns, p.maxNumBatches.at(p.job2machine.at(parent)) * p.maxNumSlotsOfJob.at(parent));
         }
-        numNeededByParent = ceilMul(parentNumRuns - max(0, maxSlots - 1), pq);
-        numNeededByParent += max(0, maxSlots - 1) * pq.toIntCeil();
+        numNeededByParent = ceilMul(parentNumRuns - max(0ll, maxSlots - 1), pq);
+        numNeededByParent += max(0ll, maxSlots - 1) * pq.toIntCeil();
       } else {
         numNeededByParent = ceilMul(parentNumRuns, pq);
       }
@@ -37,7 +37,7 @@ class Util {
     return totalNumNeeded;
   }
 
-  static int getNumRuns(int node, map<int, int>& inventory, map<int, int>& numRuns, const bool getMax, Problem& p) {
+  static int64_t getNumRuns(int node, map<int, int64_t>& inventory, map<int, int64_t>& numRuns, const bool getMax, Problem& p) {
     if (numRuns.contains(node)) {
       return numRuns.at(node);
     }
@@ -45,11 +45,11 @@ class Util {
       numRuns[node] = p.runsExcess.at(node);
       return numRuns[node];
     }
-    int numNeeded = getNumNeeded(node, inventory, numRuns, getMax, p);
+    int64_t numNeeded = getNumNeeded(node, inventory, numRuns, getMax, p);
 
     if (inventory.contains(node)) {
-      int numRemain = max(0, inventory.at(node) - numNeeded);
-      numNeeded = max(0, numNeeded - inventory.at(node));
+      int64_t numRemain = max(0ll, inventory.at(node) - numNeeded);
+      numNeeded = max(0ll, numNeeded - inventory.at(node));
       inventory[node] = numRemain;
     }
 
@@ -62,9 +62,9 @@ class Util {
     return numRuns[node];
   }
 
-  static map<int, int> getNumRunsHelper(bool getMax, Problem& p) {
-    map<int, int> inventory(p.inventory);
-    map<int, int> result;
+  static map<int, int64_t> getNumRunsHelper(bool getMax, Problem& p) {
+    map<int, int64_t> inventory(p.inventory);
+    map<int, int64_t> result;
     for (int node : p.jobTypes) {
       getNumRuns(node, inventory, result, getMax, p);
     }
@@ -72,16 +72,16 @@ class Util {
   }
 
 public:
-  static map<int, int> getMinRunsPerJob(Problem& p) { return getNumRunsHelper(false, p); }
-  static map<int, int> getMaxRunsPerJob(Problem& p) { return getNumRunsHelper(true, p); }
+  static map<int, int64_t> getMinRunsPerJob(Problem& p) { return getNumRunsHelper(false, p); }
+  static map<int, int64_t> getMaxRunsPerJob(Problem& p) { return getNumRunsHelper(true, p); }
 
-  static int ceilDiv(int x, int y) { return (x + y - 1) / y; }
-  static int ceilMul(int x, const Fraction& y) { return (x * y.num + y.den - 1) / y.den; }
-  static int roundMul(int x, const Fraction& y) { return (x * y.num + y.den / 2) / y.den; }
+  static int64_t ceilDiv(int64_t x, int64_t y) { return (x + y - 1) / y; }
+  static int64_t ceilMul(int64_t x, const Fraction& y) { return (x * y.num + y.den - 1) / y.den; }
+  static int64_t roundMul(int64_t x, const Fraction& y) { return (x * y.num + y.den / 2) / y.den; }
 
-  static int gcds(vector<int> xs) {
-    int gcd = 1;
-    for (int x : xs) {
+  static int64_t gcds(vector<int64_t> xs) {
+    int64_t gcd = 1;
+    for (int64_t x : xs) {
       gcd = std::gcd(x, gcd);
     }
     return gcd;
